completed

+++ lab11_1.cpp
@@ -1,14 +1,14 @@
 #include<iostream>
 using namespace std;
 
-int fibonacci(int);
+long long int fibonacci(int);
 
 int main(){
     cout << fibonacci(50);
     return 0; 
 }
 
-int fibonacci(int x){
+long long int fibonacci(int x){
     if(x == 0) return 0;
     else if(x == 1) return 1;
     else if(x>1){
